+++ RELEASE_NOTES.md
@@ -101,6 +101,7 @@
 - **Tables / Auto-Naming**: Fixed suggested file name suffixes for .fpt tables.
 - **Tables / Overview**: Default sorting by display name is not case-sensitive anymore.
 - **Backglass Manager / Grill visibility**: When grill visibility is standard, Backglass preview now takes in account the global grill visibility.
+- **Backglass Manager / support of sub-folders**: When table is in sub-folder, the backglass and associated res or screenres files were not properly got.
 - **Studio Exit Dialog**: Fixed issues when cancelling the exit dialog.
 - **Highscore Cards Editor**: Fixed various weight and posture issues with the font selection and the preview of it.
 - **Tables / Table Data Manager**: Fixed error during auto-applying values from VPS tables selection.